Duke A-Packages: Organize into Packages
Move DukeException class into new package ip.duke.exceptions

Import the new package into Duke.java (to use DukeException)
