+++ src/main/java/ip/duke/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package ip.duke;
+package ip.duke.exceptions;
 
 public class DukeException extends Exception {
     public DukeException(String message, Throwable ex) {
